requirements.txt
streamlit
pandas
numpy
sentence-transformers
torch
pymupdf
python-docx
scikit-learn
